SD/fix/#49 Refactored book, category, author and created tests

+++ online-book-store/src/main/java/com/amigoscode/cohort2d/onlinebookstore/book/BookController.java
@@ -20,7 +20,7 @@ public class BookController {
     }
 
     @GetMapping("{id}")
-    public BookDTO getBookById(@PathVariable("id") int id){
+    public BookDTO getBookById(@PathVariable("id") Long id){
         return bookService.getBookById(id);
     }
